Extract the VideoPlayBadge into its own file
