+++ ElementX/Sources/Screens/Timeline/View/TimelineItemViews/VideoRoomTimelineView.swift
@@ -137,18 +137,3 @@ struct VideoRoomTimelineView_Previews: PreviewProvider, TestablePreview {
                               properties: .init(isEdited: isEdited))
     }
 }
-
-/// The play badge drawn over video thumbnails, in the timeline and on the media viewer's posters.
-struct VideoPlayBadge: View {
-    var body: some View {
-        CompoundIcon(\.playSolid, size: .medium, relativeTo: .compound.headingLG)
-            .foregroundStyle(.compound.iconPrimary)
-            .padding(13)
-            .background {
-                ZStack {
-                    Circle().fill(.compound.bgSubtleSecondary)
-                    Circle().stroke(.compound.borderInteractiveSecondary)
-                }
-            }
-    }
-}
